[scroll-animations] Removing element should null out related timelines on existing animations https://bugs.webkit.org/show_bug.cgi?id=286328 rdar://143352487
Reviewed by NOBODY (OOPS!).

For the "Removing/inserting element with attaching timeline" subtest of css/timeline-scope.html,
an element with an associated named timeline is removed from the dom, which should result in the
animation attached to the associated timeline to become null. This progresses one part of the subtest,
but is still failing due to an issue where attaching to a new timeline doesn't add the associated effect
to the KeyframeEffectStack because the animation isn't considered relevant (this affects another subtest
as well).

* Source/WebCore/animation/AnimationTimelinesController.cpp:
(WebCore::AnimationTimelinesController::unregisterNamedTimelinesAssociatedWithElement):